btb-pred: integrated aBTB with BPU, bug need to be fixed.
after integrating aBTB with BPU, we encountered a bug where the Fetch Block predicted by aBTB
has branch that are outside the FB's limit. This is due to a type of Aliasing caused by two FB having different
index bits. I implemented a unit test to demonstrate this type of bug.