Cherry-pick multiple commits from mainline gem5 for VLA fix (#382)
* cpu: traffic_gen: Use heap-alloc buf instead of non-standard VLA

Allocate outside the loop to reduce the overhead of repeated heap
allocation.

Change-Id: I29b997606db8b7453c7bc696be29a84c3e1f8de6

* arch-arm: nativetrace: Use heap buf instead of VLA

Change-Id: I493c932d17cfe1c666de341b13e06674d1ef0dba

* arch-arm: gicv3 Use heap allocs instead of VLAs

This is done in the context of reading/writing LPI state to
memory, so the additional heap allocation is relatively cheap.

Change-Id: Ie1dd5604db64da05e95591e1103b62ad7e981c33

* arch-arm: curTaskNameFromTaskStruct: fix use of VLA

Instead use a fixed-size std::string

It's not performance critical, and in any event using std::string
is is in fact better than existing implementation which
implictly constructs a tmp std::string from the buffer as the
return value. (And in both cases the return value undergoes copy
elision)

Change-Id: Iee76d2aed680c7926e25e9223d06287ecfca48de

* mem-ruby: printFaultVector: Avoid using VLA

Change-Id: I0bfd029348acd0091551c6e1970607fc1c623ead

* mem: Compressor: Use heap alloc instead of VLA

Change-Id: I0c441498fcafcbf33e0e7ba7bc866f8d156a4f6f

* base: Fix failing compiler tests (#1793)

This PR adds `#include <algorithm>` to `src/base/random.cc`. This should
fix some of the failing compiler tests.

* base: cprintf: Use string for padding instead of VLAs

In practice, short string optimisation will probably apply
here, avoiding a heap allocation entirely.

Change-Id: I2cb86aa380c4fe9a00327f98e58f4c57c56a1a48

* base: RemoteGDB: Use strings instead of VLAs

Change-Id: I2d6334b8eef80bd7a28bfc311093eb72a43c2783

systemc: string for x-value print

Change-Id: I06e168db56eab81323f747b119a56f481603b696

* systemc: Use string for VcdTraceValInt fmt instead of VLA

This is actually more efficient than the existing implementation
1. The printVal call was implicitly constructing a string from str
   anyway.
2. The string will often be short enough for short string optimisation
   and will have no heap allocations.

Change-Id: Ide153d386c59a7a9005ed94bb1f48db257fd9b1f

* python: Use heap alloc for buf when unmarshalling

VLAs are non-portable, and the allocation is cheap relative
to the cost of decompression and the subsequent buffer copy
done by PyBytes_FromStringAndSize in py::bytes()

Change-Id: Ie8546aa4879d815e358d7bcfd778295f5941afef

* mem-ruby: Fix misprint in RubyRequest sequencer ostream<< (#2199)

It was printing [ k = v v[ k = v v ] ]
instead of      [ k = v v ] [ k = v v ]

Remove an unused and undefined declaration in Sequencer.hh while we're
here.

Change-Id: I7d250e40490b4840c288678979fbc1809ac592d8

* sim: replicatePage: Use heap buffer to copy page content

This only happens when doing syscall emulation of clone()
which isn't very frequent, and in any event doing page-sized
stack allocations (potentially 64kB) isn't great if we
ever want to use this in stack-constrained settings.

Change-Id: Ic0d9375141a2cacef8b1c9fb754a1ee869cce6b2

* sim: Use heap allocs for native syscall I/O

For the syscall emulation of readv, writev, poll, recvmsg.

VLAs are non-portable, and in the context of the accesses
through the simulated memory hierachy, the additional heap
allocations don't cost much.

Change-Id: I972c68fe4dd84a17f0ab0feb8eb17ad3bce545aa

* dev-virtio: Use heap alloc when read/writing virtio ring

Instead of a ISO C++ VLA. There's an extra heap allocation here
but it's inexpensive relative to the cost of copying the entire
virtio ring structure from simulated memory and adjusting the
endianness.

Change-Id: Ia2e6d5ad711d0c402e8e84c57876d8b402184917

* dev-virtio: Use heap alloc when forwarding ring data to console

Instead of a non ISO C++ VLA.

The cost of a single heap allocation will be low relative to the
cost of a write() syscall or ostream::put for each character output.

Change-Id: I911e962d0f378b0af341752c96b7616adb108474

* dev-virtio: fs9p: Use heap allocs instead of VLAs

VLAs are not portable ISO C++

NB in recvTMsg we explicitly use malloc instead of make_unique to
avoid a GCC 14 memcpy bounds check warning because it can't detect
the size of the buffer properly.

Change-Id: I9935ae115885c53d6120248dce9359733fd4e61d

* arch: addr_range: Use heap allocs instead of VLAs (#2206)

This is only used by getOffset with interleaved AddrRanges and it
already has 2 std::sort invocations in it, so an additional short-lived
heap allocation shouldn't be too much of a concern.

* arch-x86: intelmp: Don't use VLA to build (string) table entries

Change-Id: I11ba5a4ea7b3739984e5418be2c7011817b1883b

* arch-x86: se_workload: Don't use VLAs when reading irq stack

Change-Id: I9c1311b097ef4674aea132a639e637dab2ba0514

* arch-power: Avoid incidental use of VLA in sizeof expression. (#2236)

Change-Id: I81db4d909aad708680df0cf1bc687175f059ae0d

* arch-mips: Avoid incidental use of VLA in sizeof expression (#2237)

Change-Id: I8a8f0fb81f28879a2d7cc1c9a317d99ffc880769

* tests: Fix compiler tests (#1678)

- This change updates syntax of constructors of Template Classes from
`class<T>()` to `class()`

- Initializes coherence to 0 in `src/mem/cache_blk.hh`

The above changes are made to solve the errors when compiling gem5 in
gcc 14

* cpu: use heap alloc instead of VLA

Change-Id: Ibf3d0730b4bf8c5106e2a6e93c83dff18680f825

* cpu,mem: fix gcc14 compile error

Change-Id: I65c0e7fb708fef95f581c14fb0e2f8b20fd7f5f8

* misc: remove clang warning disabling

now can be compiled under gcc14, clang18, clang19
no VLAs in common code

Change-Id: Ibdd805bc613ff283820d52e12a1619e1713d7e51

---------

Co-authored-by: Brian Foley <bpfoley@gmail.com>
Co-authored-by: Ivana Mitrovic <imitrovic@ucdavis.edu>
Co-authored-by: Harshil Patel <hpppatel@ucdavis.edu>