cpu: fix LSQRequest num control bugs by moving into LSQUnit (#500)
During multi-core operation, we observed a situation where the number of numSBufferRequests exceeded the upper limit of port->sbufferEntries. This is because numSBufferRequest was wrongly defined as a static variable, which resulted in different lsqUnits sharing the same numSBufferRequest. As a result, the numSBufferRequest in each lsqUnit did not exceed its respective sbufferEntries, but the total exceeded the limit. Therefore, it is modified to a non-static variable and moved inside the LSQUnit. Each lsqUnit maintains its own set of values, and numSingleRequest and numSplitRequest follow the same pattern.

Change-Id: I42e41a69333ce85d3145cae13441eda0ab3cc158